feat: data quality + refactoring
