+++ spark/jobs/current_job.py
@@ -9,7 +9,7 @@ from pyspark.sql.types import StructType, StructField, StringType
 from jobs.metrics.statistics import calculate_statistics_current
 from jobs.models.current_dataset import CurrentDataset
 from jobs.models.reference_dataset import ReferenceDataset
-from utils.current import CurrentMetricsService
+from utils.current_binary import CurrentMetricsService
 from utils.models import JobStatus, ModelOut
 from utils.db import update_job_status, write_to_db
