+++ spark/jobs/utils/spark.py
@@ -1,7 +1,11 @@
-from pyspark.sql.functions import col
+import pyspark.sql.functions as F
 
 
 def apply_schema_to_dataframe(df, schema):
     for field in schema.fields:
-        df = df.withColumn(field.name, col(field.name).cast(field.dataType))
+        df = df.withColumn(field.name, F.col(field.name).cast(field.dataType))
     return df
+
+
+def check_not_null(x):
+    return F.when(F.col(x).isNotNull() & ~F.isnan(x), F.col(x))
