+++ spark/tests/multiclass_reference_test.py
@@ -1,6 +1,7 @@
 import datetime
 import uuid
 
+import deepdiff
 import pytest
 
 from jobs.metrics.statistics import calculate_statistics_reference
@@ -122,9 +123,8 @@ def test_calculation_dataset_target_string(spark_fixture, dataset_target_string)
 
     multiclass_service = ReferenceMetricsMulticlassService(reference_dataset)
 
-    print(multiclass_service.calculate_data_quality().model_dump_json(serialize_as_any=True))
-
     stats = calculate_statistics_reference(reference_dataset)
+    data_quality = multiclass_service.calculate_data_quality()
 
     assert stats == my_approx(
         {
@@ -138,4 +138,100 @@ def test_calculation_dataset_target_string(spark_fixture, dataset_target_string)
             "n_variables": 7,
             "numeric": 2,
         }
+    )
+
+    assert not deepdiff.DeepDiff(
+        data_quality.model_dump(serialize_as_any=True, exclude_none=True),
+        {
+            "n_observations": 10,
+            "class_metrics": [
+                {"name": "ORPHAN", "count": 2, "percentage": 20.0},
+                {"name": "UNHEALTHY", "count": 2, "percentage": 20.0},
+                {"name": "HEALTY", "count": 4, "percentage": 40.0},
+                {"name": "UNKNOWN", "count": 2, "percentage": 20.0},
+            ],
+            "feature_metrics": [
+                {
+                    "feature_name": "num1",
+                    "type": "numerical",
+                    "missing_value": {"count": 1, "percentage": 10.0},
+                    "mean": 1.1666666666666667,
+                    "std": 0.75,
+                    "min": 0.5,
+                    "max": 3.0,
+                    "median_metrics": {"perc_25": 1.0, "median": 1.0, "perc_75": 1.0},
+                    "class_median_metrics": [],
+                    "histogram": {
+                        "buckets": [
+                            0.5,
+                            0.75,
+                            1.0,
+                            1.25,
+                            1.5,
+                            1.75,
+                            2.0,
+                            2.25,
+                            2.5,
+                            2.75,
+                            3.0,
+                        ],
+                        "reference_values": [2, 0, 5, 0, 1, 0, 0, 0, 0, 1],
+                    },
+                },
+                {
+                    "feature_name": "num2",
+                    "type": "numerical",
+                    "missing_value": {"count": 2, "percentage": 20.0},
+                    "mean": 277.675,
+                    "std": 201.88635947695215,
+                    "min": 1.4,
+                    "max": 499.0,
+                    "median_metrics": {
+                        "perc_25": 117.25,
+                        "median": 250.0,
+                        "perc_75": 499.0,
+                    },
+                    "class_median_metrics": [],
+                    "histogram": {
+                        "buckets": [
+                            1.4,
+                            51.160000000000004,
+                            100.92000000000002,
+                            150.68000000000004,
+                            200.44000000000003,
+                            250.20000000000002,
+                            299.96000000000004,
+                            349.72,
+                            399.48,
+                            449.24,
+                            499.0,
+                        ],
+                        "reference_values": [1, 1, 1, 1, 0, 0, 1, 0, 0, 3],
+                    },
+                },
+                {
+                    "feature_name": "cat1",
+                    "type": "categorical",
+                    "missing_value": {"count": 0, "percentage": 0.0},
+                    "category_frequency": [
+                        {"name": "B", "count": 4, "frequency": 0.4},
+                        {"name": "C", "count": 1, "frequency": 0.1},
+                        {"name": "A", "count": 5, "frequency": 0.5},
+                    ],
+                    "distinct_value": 3,
+                },
+                {
+                    "feature_name": "cat2",
+                    "type": "categorical",
+                    "missing_value": {"count": 0, "percentage": 0.0},
+                    "category_frequency": [
+                        {"name": "Y", "count": 1, "frequency": 0.1},
+                        {"name": "X", "count": 9, "frequency": 0.9},
+                    ],
+                    "distinct_value": 2,
+                },
+            ],
+        },
+        ignore_order=True,
+        significant_digits=6,
     )
